backend/requirements.txt
# FastAPI Backend Dependencies

# Core FastAPI stack
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database - MongoDB
motor==3.4.0
beanie==1.26.0
pymongo==4.6.1

# HTTP client
httpx==0.27.0

# Configuration
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.1.1


# Development tools
pytest==8.3.4
black==24.10.0
flake8==7.1.1

# Logging
structlog==23.2.0

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1